Refactor Dework code to new code structure
