+++ src/lib/gql/mutations.ts
@@ -65,3 +65,31 @@ export const updateCircleLogo = async (
       },
     }
   );
+
+export const createCircleIntegration = async (
+  circleId: number,
+  type: string,
+  name: string,
+  data: any
+) =>
+  client.mutate(
+    {
+      insert_circle_integrations_one: [
+        {
+          object: {
+            circle_id: circleId,
+            type,
+            name,
+            data: $`data`,
+          },
+        },
+        { id: true },
+      ],
+    },
+    { variables: { data } }
+  );
+
+export const deleteCircleIntegration = async (id: number) =>
+  client.mutate({
+    delete_circle_integrations_by_pk: [{ id }, { id: true }],
+  });
